pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.1",
  "python-telegram-bot==13.15",
  "telegram>=0.0.1",
]
